Restructured folders
All datasets now have their own folder that will include scripts, requirements file, and a data folder.
All data files for each set will now be kept in the data folder, and not uploaded to github.
May be beneficial to include a data-link file in each set, to allow easy downloading of the data.

Also moved the example scripts and associated datafiles to an '_example_scripts' folder.